disable cacheRefs
The purpose they serve is to update all other references to a
specific resource id (eg a user) if some other (newer) request has that resource
as one of the results.

The problem is resources have different structures depending on if it's a
private/public reference (eg getMe vs get list of users where I'm in that list)
or if some data is embedded (eg embed recent games played).

The cache needs to reflect that, but it doesn't look like an easy fix